feat: Level 8 완료

+++ src/main/java/jscode/board/service/BoardService.java
@@ -1,14 +1,21 @@
 package jscode.board.service;
 
 import jscode.board.domain.Board;
+import jscode.board.domain.Comment;
+import jscode.board.domain.Member;
 import jscode.board.dto.board.BoardRequestDto;
 import jscode.board.dto.board.BoardResponseDto;
 import jscode.board.exception.board.BoardNotFoundException;
+import jscode.board.exception.member.NotFoundMemberException;
 import jscode.board.repository.BoardRepository;
+import jscode.board.repository.CommentRepository;
+import jscode.board.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,10 +25,13 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BoardService {
     private final BoardRepository boardRepository;
+    private final MemberRepository memberRepository;
+    private final CommentRepository commentRepository;
 
     @Transactional
     public BoardResponseDto createBoard(BoardRequestDto req) {
-        Board board = boardRepository.save(req.toEntity());
+        Member member = memberRepository.findById(jwtAuth()).orElseThrow(NotFoundMemberException::new);
+        Board board = boardRepository.save(new Board(req.getHead(), req.getBody(), member));
         BoardResponseDto boardResponseDto = BoardResponseDto.toDto(board);
         return boardResponseDto;
     }
@@ -44,6 +54,9 @@ public class BoardService {
     @Transactional
     public BoardResponseDto editBoard(Long id, BoardRequestDto req) {
         Board board = boardRepository.findById(id).orElseThrow(BoardNotFoundException::new);
+        if (board.getMember().getId() != jwtAuth()) {
+            throw new RuntimeException("게시글 유저 키와 jwt 유저 키가 일치하지 않습니다.");
+        }
         board.editBoard(req);
         return BoardResponseDto.toDto(board);
     }
@@ -51,6 +64,9 @@ public class BoardService {
     @Transactional
     public void deleteBoard(Long id) {
         Board board = boardRepository.findById(id).orElseThrow(BoardNotFoundException::new);
+        if (board.getMember().getId() != jwtAuth()) {
+            throw new RuntimeException("게시글 유저 키와 jwt 유저 키가 일치하지 않습니다.");
+        }
         boardRepository.delete(board);
     }
 
@@ -60,6 +76,13 @@ public class BoardService {
         List<Board> results = boardRepository.findByHeadContaining(head, pageable);
         results.forEach(e -> BoardResponseDto.toDto(e));
         return results;
+    }
+
+    public Long jwtAuth() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String name = authentication.getName();
+        Long id = Long.parseLong(name);
+        return id;
     }
 
 }
